Driving and walking directions from Cartociudad API

Cartociudad API provides driving and walking routes between two points. This function quieries the API and provides the user the data in convenient form.

cartociudad_get_route(latlon.orig, latlon.dest, vehicle = "car")

Arguments

latlon.orig
Latitude and longitude of the starting point

latlon.dest
Latitude and longitude of the destination point

vehicle
Either car or walking

Value

A list containing the fields described in Cartociudad API documentation (see the link below).

References

http://www.cartociudad.es/recursos/Documentacion_tecnica/CARTOCIUDAD_ServiciosWeb.pdf

Author

Carlos J. Gil Bellosta

Examples

res <- cartociudad_get_route(c(39.48,-0.37),
   c(39.484336,-0.358171),
   vehicle = "car")

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.